Write point texture indices as bytes

+++ src/sysdolphin/baselib/psdisp.c
@@ -630,7 +630,7 @@ static inline HSD_Particle* psDispSubPoint(HSD_Particle* pp)
             GXWGFifo.f32 = y;
             GXWGFifo.f32 = z;
             if (pp->kind & DispTexture) {
-                GXWGFifo.f32 = 1;
+                GXWGFifo.u8 = 1;
             }
         }
     }
@@ -1107,7 +1107,6 @@ static inline void psDispSubMakePolygon(HSD_Particle* pp, u8* texform, f32 x,
     } else {
         u8* it = texform;
         u32 primitive_count = *(u32*) it;
-
         it += sizeof(u32);
         for (; primitive_count != 0; primitive_count--) {
             GXPrimitive primitive = it[0];
@@ -1948,6 +1947,7 @@ static inline void psDispSubAppSRT(HSD_Particle* pp, u8* texform,
     } else {
         u8* it = texform;
         u32 primitive_count = *(u32*) it;
+
         it += sizeof(u32);
         for (; primitive_count != 0; primitive_count--) {
             GXPrimitive prim = it[0];
